NODE-667 State synthetic benchmarks: one file for all benchmarks
